Connect to devices without delay
The reconnect function in tuydriver was doing reconnects every
20 seconds, but this function was also used for the initial
connect on app startup, or when a new device was added,
resulting in the device being unavailable for longer
than necessary. This was observed in #35.

We now do an immediate connect on device init, and
then spin off a repeating timed reconnect every 20
seconds. By decoupling the connect logic from the
Homey device state, we also ensure that we reconnect
regardless of whether the plumbing to Homey has been
made aware of the disconnect.

The existing tuydriver.reconnect() has been left as
is, for the deprecated drivers.